+++ src/com/sproutlife/model/rotations/Rotation.java
@@ -5,7 +5,7 @@
  * The above copyright notice and this permission notice shall be included in all
  * copies or substantial portions of the Software.
  *******************************************************************************/
-package com.sproutlife.geometry;
+package com.sproutlife.model.rotations;
 
 /**
  * @author Alex Shapiro
