+++ src/com/sproutlife/model/seed/Seed.java
@@ -9,9 +9,9 @@ package com.sproutlife.model.seed;
 
 import java.awt.Point;
 
-import com.sproutlife.geometry.Rotation;
-import com.sproutlife.geometry.Rotations;
 import com.sproutlife.model.echosystem.Organism;
+import com.sproutlife.model.rotations.Rotation;
+import com.sproutlife.model.rotations.Rotations;
 
 public class Seed {
